requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
deepseek-sdk==0.1.0
distro==1.9.0
gigachat==0.1.42.post2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.75
langchain-gigachat==0.3.12
langsmith==0.4.23
openai==1.104.2
orjson==3.11.3
packaging==25.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
Telethon==1.41.0
tenacity==9.1.2
tqdm==4.67.1
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
zstandard==0.24.0
psutil==5.9.8
